Serial Com test program
The program will allow for two way communication with another MCU
through the use of the Serial1 port. Arduino must be connected to PC in
order to send the messages. The MCUs can either be connected directly to
each other by wires or through the use of the bluetooth board and
bluetooth modules. The modules need to be paired beforehand via PC.
Currently, communication via bluetooth is possible, but there is an
issue where the first part of a message gets jumbled and makes no sense
which happens every time. Need to investigate further.