Created master code
Started work on the master code. Established a core version. Get's
distance measurement from potentiometer. Cannot currently get the speed
so a test value of 80 km/h is used in the code. Also able to receive
commands from Android Application. Currently, the only command available
is "D" which returns the distance of the car behind user to the app and
also prints it to the consol.