Added ultrasonic sensor code, new commands
Completed code for ultrasonic sensor. The pot is no longer needed and
the position sensing works (though scaled down). New "T" command
terminates the BT connection. Work on "P" command started but not yet
completed, uses EEPROM to store the pin for the BT module which is where
it will be stored and changed.